requirements.txt
Django~=4.1.3
djangorestframework==3.14.0
django-environ~=0.9.0
django-filter~=22.1
psycopg2-binary~=2.9.5
gunicorn~=20.1.0
pandas~=1.5.3
openpyxl~=3.1.1